fix / update using the newest XMP*StaticRelease64.lib s that I built separately

+++ SE.Halligang/CsXmpToolkit/CsXmpToolkitNetStandard/Enums/PInvoke/TimeZoneSign.cs
@@ -1,6 +1,6 @@
 namespace SE.Halligang.CsXmpToolkit.PInvoke
 {
-	internal enum TimeZoneSign : int
+	internal enum TimeZoneSign : sbyte
 	{
 		WestOfUtc	= -1,
 		IsUtc		= 0,
